Moved ImmutableBuffer to a separate file.
